Split initializers and classes
