+++ app/javascript/active_admin/initializers/batch-actions.js
