+++ app/javascript/active_admin/initializers/checkbox-toggler.js
@@ -1,0 +1,3 @@
+import CheckboxToggler from "../lib/checkbox-toggler";
+
+$.widget.bridge('checkboxToggler', CheckboxToggler);
