+++ app/javascript/active_admin/initializers/dropdown-menu.js
@@ -1,0 +1,9 @@
+import DropdownMenu from "../lib/dropdown-menu";
+
+$.widget.bridge('aaDropdownMenu', DropdownMenu);
+
+const onDOMReady = () => $('.dropdown_menu').aaDropdownMenu();
+
+$(document).
+  ready(onDOMReady).
+  on('page:load turbolinks:load', onDOMReady);
