+++ app/javascript/active_admin/initializers/has-many.js
