+++ app/javascript/active_admin/initializers/per-page.js
@@ -1,0 +1,13 @@
+import PerPage from "../lib/per-page";
+
+(($) => {
+
+  $(document).
+    on('change', '.pagination_per_page > select', function(event) {
+      PerPage._jQueryInterface.call($(this), 'update')
+    });
+
+  $.fn['perPage'] = PerPage._jQueryInterface
+  $.fn['perPage'].Constructor = PerPage
+
+})(jQuery);
