+++ app/javascript/active_admin/initializers/table-checkbox-toggler.js
@@ -1,0 +1,3 @@
+import TableCheckboxToggler from "../lib/table-checkbox-toggler";
+
+$.widget.bridge('tableCheckboxToggler', TableCheckboxToggler);
