+++ app/javascript/active_admin/lib/per-page.js
@@ -1,48 +1,38 @@
 import { queryStringToParams, hasTurbolinks, turbolinksVisit, toQueryString } from './utils';
 
-(($) => {
+class PerPage {
+  constructor(element) {
+    this.element = element;
+  }
 
-  class PerPage {
-    constructor(element) {
-      this.element = element;
+  update() {
+    const params = queryStringToParams()
+      .filter(({name}) => name != 'per_page' || name != 'page')
+
+    params.push({ name: 'per_page', value: this.element.value });
+
+    if (hasTurbolinks()) {
+      turbolinksVisit(params);
+    } else {
+      window.location.search = toQueryString(params);
     }
+  }
 
-    update() {
-      const params = queryStringToParams()
-        .filter(({name}) => name != 'per_page' || name != 'page')
+  static _jQueryInterface(config) {
+    return this.each(function () {
+      const $this = $(this)
+      let data = $this.data('perPage')
 
-      params.push({ name: 'per_page', value: this.element.value });
-
-      if (hasTurbolinks()) {
-        turbolinksVisit(params);
-      } else {
-        window.location.search = toQueryString(params);
+      if (!data) {
+        data = new PerPage(this)
+        $this.data('perPage', data)
       }
-    }
 
-    static _jQueryInterface(config) {
-      return this.each(function () {
-        const $this = $(this)
-        let data = $this.data('perPage')
+      if (config === 'update') {
+        data[config]()
+      }
+    })
+  }
+};
 
-        if (!data) {
-          data = new PerPage(this)
-          $this.data('perPage', data)
-        }
-
-        if (config === 'update') {
-          data[config]()
-        }
-      })
-    }
-  };
-
-  $(document).
-    on('change', '.pagination_per_page > select', function(event) {
-      PerPage._jQueryInterface.call($(this), 'update')
-    });
-
-  $.fn['perPage'] = PerPage._jQueryInterface
-  $.fn['perPage'].Constructor = PerPage
-
-})(jQuery);
+export default PerPage;
